Initial code for docker code runner
